Changed order timestamp from float to int

+++ Tribler/Core/Modules/restapi/market/asks_bids_endpoint.py
@@ -129,7 +129,7 @@ class AsksEndpoint(BaseAsksBidsEndpoint):
             if not request.finished:
                 request.write(json.dumps({
                     'assets': ask.assets.to_dictionary(),
-                    'timestamp': float(ask.timestamp),
+                    'timestamp': int(ask.timestamp),
                     'trader_id': ask.order_id.trader_id.as_hex(),
                     'order_number': int(ask.order_id.order_number),
                     'timeout': int(ask.timeout)
@@ -236,7 +236,7 @@ class BidsEndpoint(BaseAsksBidsEndpoint):
             if not request.finished:
                 request.write(json.dumps({
                     'assets': bid.assets.to_dictionary(),
-                    'timestamp': float(bid.timestamp),
+                    'timestamp': int(bid.timestamp),
                     'trader_id': bid.order_id.trader_id.as_hex(),
                     'order_number': int(bid.order_id.order_number),
                     'timeout': int(bid.timeout)
