+++ Tribler/community/market/core/order.py
@@ -169,11 +169,11 @@ class Order(object):
         Returns a database representation of an Order object.
         :rtype: tuple
         """
-        completed_timestamp = float(self.completed_timestamp) if self.completed_timestamp else None
+        completed_timestamp = int(self.completed_timestamp) if self.completed_timestamp else None
         return (database_blob(bytes(self.order_id.trader_id)), text_type(self.order_id.order_number),
                 self.assets.first.amount, text_type(self.assets.first.asset_id), self.assets.second.amount,
                 text_type(self.assets.second.asset_id), self.traded_quantity, int(self.timeout),
-                float(self.timestamp), completed_timestamp, self.is_ask(), self._cancelled, self._verified)
+                int(self.timestamp), completed_timestamp, self.is_ask(), self._cancelled, self._verified)
 
     @property
     def reserved_ticks(self):
@@ -405,7 +405,7 @@ class Order(object):
         """
         Return a dictionary representation of this order.
         """
-        completed_timestamp = float(self.completed_timestamp) if self.completed_timestamp else None
+        completed_timestamp = int(self.completed_timestamp) if self.completed_timestamp else None
         return {
             "trader_id": self.order_id.trader_id.as_hex(),
             "order_number": int(self.order_id.order_number),
@@ -413,7 +413,7 @@ class Order(object):
             "reserved_quantity": self.reserved_quantity,
             "traded": self.traded_quantity,
             "timeout": int(self.timeout),
-            "timestamp": float(self.timestamp),
+            "timestamp": int(self.timestamp),
             "completed_timestamp": completed_timestamp,
             "is_ask": self.is_ask(),
             "cancelled": self.cancelled,
@@ -430,5 +430,5 @@ class Order(object):
             "assets": self.assets.to_dictionary(),
             "traded": self.traded_quantity,
             "timeout": int(self.timeout),
-            "timestamp": float(self.timestamp)
+            "timestamp": int(self.timestamp)
         }
