+++ Tribler/community/market/core/payment.py
@@ -35,7 +35,7 @@ class Payment(Message):
         transaction_id = TransactionId(TraderId(bytes(transaction_trader_id)), TransactionNumber(transaction_number))
         return cls(TraderId(bytes(trader_id)), transaction_id, AssetAmount(transferred_amount, str(transferred_id)),
                    WalletAddress(str(address_from)), WalletAddress(str(address_to)), PaymentId(str(payment_id)),
-                   Timestamp(float(timestamp)), bool(success))
+                   Timestamp(timestamp), bool(success))
 
     def to_database(self):
         """
@@ -45,7 +45,7 @@ class Payment(Message):
         return (database_blob(bytes(self.trader_id)), database_blob(bytes(self.transaction_id.trader_id)),
                 int(self.transaction_id.transaction_number), text_type(self.payment_id), self.transferred_assets.amount,
                 text_type(self.transferred_assets.asset_id), text_type(self.address_from),
-                text_type(self.address_to), float(self.timestamp), self.success)
+                text_type(self.address_to), int(self.timestamp), self.success)
 
     @property
     def transaction_id(self):
@@ -114,6 +114,6 @@ class Payment(Message):
             "payment_id": str(self.payment_id),
             "address_from": str(self.address_from),
             "address_to": str(self.address_to),
-            "timestamp": float(self.timestamp),
+            "timestamp": int(self.timestamp),
             "success": self.success
         }
